requirements.txt
aniso8601==7.0.0
attrs==19.3.0
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
commitizen==1.16.2
coverage==5.0.3
decli==0.5.1
dnspython==1.16.0
firebase-admin==3.2.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-GraphQL==2.0.1
Flask-Limiter==1.1.0
flask-restplus==0.13.0
Flask-SeaSurf==0.2.2
flask-talisman==0.7.0
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gql==0.3.0
graphene==2.1.8
graphene-mongo==0.2.9
graphql-core==2.3.1
graphql-relay==2.0.1
graphql-server-core==1.2.0
grpcio==1.27.1
gunicorn==20.0.4
httplib2==0.17.0
idna==2.8
importlib-metadata==1.5.0
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
kubernetes==10.0.1
limits==1.5
MarkupSafe==1.1.1
mongoengine==0.19.1
mongomock==3.19.0
more-itertools==8.2.0
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.1
pluggy==0.13.1
promise==2.3
prompt-toolkit==3.0.3
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.10.1
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.3.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
questionary==1.5.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
Rx==1.6.1
sentinels==1.0.0
singledispatch==3.4.0.3
six==1.14.0
stripe==2.42.0
termcolor==1.1.0
tomlkit==0.5.8
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
websocket-client==0.57.0
Werkzeug==0.16.1
zipp==2.2.0